Add argument resolver infrastructure for CoffeeNetProfile
In the previous CoffeeNet iteration a profile for the currently logged
in user could be retrieved via an exposed service. Most of the time this
feature is needed for frontend work. Therefore we implement this
reworked implementation via the Spring mechanism of a
`HandlerMethodArgumentResolver`.

This commit adds the needed auto-configuration as also an implementation
for OAuth2/OIDC.