Initial rewrite of build from Maven to Gradle
Spring Boot moved to Gradle for its build for the upcoming 2.3.x
release. Since CoffeeNet is a micro-framework on top of Spring Boot it
should try to be aligned with the upstream framework. This commit is an
initial effort, to create the artifacts from a Gradle build, making use
of its powerful dependency mechanisms. The following parts of the
previous Maven based build are already addressed:

* Run a complex multi-project (multi-module in Maven) for the different
  parts of the framework.
* Enforce Java code convetions, forcing a 1.8 compatibility
* Generate Maven-compatible artifacts
  * … a BOM in coffeenet-dependencies
  * … an optional-enabled coffeenet-autoconfigure POM
  * … POMs with CoffeeNet organization information
  * … Sonatype OSS compatible POMs
  * … Javadoc and source artifacts
  * … a coffeenet-starter-parent for the Spring Boot version the build
    was executed against.
* Generate Sonarqube report output from tests

Additionally it adds the following features:

* Publishing of Gradle Metadata for use in Gradle based projects
* Add Java 9+ metadata to built JARs with automatic module information
* Shorten the build descriptors for the individual modules to around
  15 to 30 lines
* Support incremental builds (via Gradle)
* Fix(!) previously flattened POMs with wrong expansion settings