[coffeenet-platform-discovery] Initial implementation
This is based on the original implementation in the legacy CoffeeNet. It
uses a Spring Cloud based `DiscoveryClient` and is not tied to a
particular implementation like Eureka.

This way it can be used with either Eureka or with a statically
configured `SimpleDiscoveryClient`. Theoretically it could also use the
Kubernetes based version, but this one does not really work for services
exposed to the outside world.